SIG-3150 refactored Sigmax PDF image handling
