+++ api/app/signals/apps/api/templates/api/pdf/print_signal.html
@@ -185,8 +185,8 @@
             <p><img src="{{ image.file.url }}" style="width:680px" alt=""></p>
         {% endfor %}
         <br>
-    {% elif jpg_data_urls %}  {# HOTFIX SIG-1473 #}
-        {% for data_url in jpg_data_urls %}
+    {% elif jpg_data_uris %}  {# HOTFIX SIG-1473 #}
+        {% for data_url in jpg_data_uris %}
             {% if data_url %}
                 <p><img src="{{ data_url|safe }}" style="width:680px" alt=""></p>
             {% else %}
